+++ SchoolSystem/SchoolSystem.Repository/Services/ITeacherRepository.cs
@@ -4,11 +4,11 @@ namespace SchoolSystem.Repository.Services;
 
 public interface ITeacherRepository
 {
-    Task<Teacher> SelectTeacherByIdAsync(int id);
+    Task<Teacher> SelectTeacherByIdAsync(long id);
     Task<List<Teacher>> SelectAllTeachersAsync(bool includeStudent, bool includeClass, int skip, int take);
     Task InsertTeacherAsync(Teacher teacher);
     Task UpdateTeacherAsync(Teacher teacher);
-    Task DeleteTeacherAsync(int id);
+    Task DeleteTeacherAsync(long id);
 }
 
 /***
